% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Calculo_metricas.R
\name{calcular_auc}
\alias{calcular_auc}
\title{Funci?n para calcular el ?rea Bajo la Curva ROC (AUC), en variables continuas (vector num?rico)}
\usage{
calcular_auc(predicciones, clase_verdadera)
}
\arguments{
\item{predicciones}{Un vector num?rico que contiene las predicciones del modelo.}

\item{clase_verdadera}{Un vector num?rico que contiene las clases verdaderas #'correspondientes a las predicciones.  Debe ser un vector binario y representa las #'clases verdaderas  (0 para negativo, 1 para positivo).}
}
\value{
Un valor num?rico que representa el AUC calculado a partir de las predicciones y las clases verdaderas.  Un n?mero en el rango de 0 a 1, donde un valor m?s alto  indica un mejor rendimiento del modelo y 0.5 indica un modelo que no es mejor que el azar. Si la variable de clase no es binaria (no tiene exactamente dos valores ?nicos), la funci?n devuelve NA y muestra una advertencia.
}
\description{
Esta funci?n calcula el AUC (?rea Bajo la Curva ROC) para evaluar
el rendimiento de un modelo de clasificaci?n binaria. El AUC mide la capacidad
del modelo para distinguir entre clases positivas y negativas. Por tanto, para distinguir #'entre clase 0 y 1.
}
\examples{
# Ejemplo de uso
predicciones <- c(0.9, 0.7, 0.6, 0.4, 0.2)
clase_verdadera <- c(1, 0, 1, 0, 1)
auc <- calcular_auc(predicciones, clase_verdadera)# Calcula el AUC del modelo.
auc

# Ejemplo con advertencia
clase_no_binaria <- c(0, 1, 2, 1, 0)
auc <-  calcular_auc(predicciones, clase_no_binaria)  # Devuelve NA con advertencia.
auc

}
