% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Correlacion_informacion_mutua.R
\name{calcular_correlacion_informacion_mutua}
\alias{calcular_correlacion_informacion_mutua}
\title{Funci?n para calcular correlaci?n o informaci?n mutua entre variables en un dataset}
\usage{
calcular_correlacion_informacion_mutua(dataset)
}
\arguments{
\item{dataset}{El dataset que contiene las variables a analizar.}
}
\value{
Una matriz de correlaciones que muestra las correlaciones (o informaci?n mutua) entre las variables del dataset. Las filas y columnas de la matriz est?n etiquetadas con los nombres de las variables originales.
}
\description{
Esta funci?n toma un dataset y calcula las correlaciones (en el caso de variables num?ricas) o la informaci?n mutua (en el caso de variables categ?ricas) entre todas las pares de variables del dataset. La funci?n devuelve una matriz de correlaciones donde cada celda representa la correlaci?n entre dos variables. Para variables incompatibles (una num?rica y una categ?rica), se asigna NA.
}
\examples{
Ejemplo de uso
dataset <- data.frame(
Variable1 = c(1.8, 2.6, 4.5, 4.0, 1.2),
Variable2 = c(0.7, 2.0, 3.5, 6.0, 0.5),
Variable3 = factor(c("B", "B", "A", "C", "C")),
Variable4 = factor(c("B", "Y", "A", "Z", "Z"))
)
matriz_correlacion <- calcular_correlacion_informacion_mutua(dataset)
matriz_correlacion

}
