% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Calculo_metricas.R
\name{calcular_entropia}
\alias{calcular_entropia}
\title{Funci?n para calcular la entrop?a de un vector discreto/factor}
\usage{
calcular_entropia(vector_discreto)
}
\arguments{
\item{vector_discreto}{Un vector discreto o factor del cual se calcular? la entrop?a. Contiene categor?as o etiquetas.}
}
\value{
Un valor num?rico que representa la entrop?a del vector discreto o factor. La entrop?a es un n?mero positivo o cero, puede estar en el rango de 0 (sin incertidumbre) a 1 (m?xima incertidumbre).
}
\description{
Esta funci?n calcula la entrop?a de un vector discreto o factor, que es una medida de la incertidumbre o desorden en las categor?as del vector. La entrop?a es una medida de la informaci?n contenida en el vector.
}
\examples{
categorias <- factor(c("A", "B", "A", "C", "B", "A"))
entropia <- calcular_entropia(categorias)
entropia

}
