% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Correlacion_informacion_mutua.R
\name{calcular_informacion_mutua}
\alias{calcular_informacion_mutua}
\title{Funci?n para calcular la informaci?n mutua entre dos variables categ?ricas}
\usage{
calcular_informacion_mutua(variable1, variable2)
}
\arguments{
\item{variable1}{Un vector categ?rico que representa la primera variable.}

\item{variable2}{Un vector categ?rico que representa la segunda variable.}
}
\value{
El valor de la informaci?n mutua entre las dos variables. Si los vectores no tienen la misma longitud, la funci?n muestra un mensaje de error y se detiene.
}
\description{
Esta funci?n calcula la informaci?n mutua entre dos variables categ?ricas, que mide la dependencia entre ellas. La informaci?n mutua cuantifica cu?nta informaci?n compartida existe entre las dos variables. Es un valor no negativo, donde 0 indica independencia y valores mayores indican mayor dependencia.
}
\examples{
Ejemplo de uso
var1 <- c("A", "B", "A", "C", "B")
var2 <- c("X", "Y", "X", "Z", "X")
calcular_informacion_mutua(var1, var2)  # Calcula la informaci?n mutua entre las dos variables.

Ejemplo con vectores de diferente longitud
var1 <- c("A", "B", "A")
var2 <- c("X", "Y", "X", "Z")
calcular_informacion_mutua(var1, var2)  # Muestra un mensaje de error.

}
