% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Calculo_metricas.R
\name{calcular_metricas}
\alias{calcular_metricas}
\title{Funci?n principal para calcular m?tricas en funci?n del tipo de atributo}
\usage{
calcular_metricas(dataset, variable_clase = NULL)
}
\arguments{
\item{dataset}{El conjunto de datos en forma de data frame.}

\item{variable_clase}{(Opcional) La variable de clase utilizada para calcular el AUC.
Un vector que contiene las clases verdaderas para calcular el AUC en caso de que los atributos sean num?ricos. Debe ser un vector binario (0/1). Por defecto, es NULL y no se calcula el AUC.  Si decides proporcionar este vector, debe ser un "factor".}
}
\value{
Un data frame que contiene m?tricas calculadas para cada atributo en el conjunto de datos, incluyendo varianza, entrop?a y AUC (si es aplicable).
}
\description{
Esta funci?n calcula m?tricas espec?ficas en funci?n del tipo de atributo en un conjunto de datos. Puede calcular varianza, entrop?a y AUC (?rea Bajo la Curva ROC) dependiendo de si los atributos son num?ricos o discretos.
}
\examples{
# Crear un conjunto de datos de ejemplo
datos <- data.frame(
  Atributo1 = c(1.2, 2.3, 3.1, 4.0, 5.5),
  Atributo2 = factor(c("A", "B", "A", "C", "B")),
  Clase = c(1, 0, 1, 0, 1)
)
resultados <- calcular_metricas(datos, datos$Clase)
resultados

}
