% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Calculo_metricas.R
\name{calcular_varianza}
\alias{calcular_varianza}
\title{Funci?n para el c?lculo de la varianza en una variable continua (vector num?rico)}
\usage{
calcular_varianza(vector_numerico)
}
\arguments{
\item{vector_numerico}{Un vector num?rico del cual se calcular? la varianza.}
}
\value{
Un valor num?rico que representa la varianza del vector proporcionado.
}
\description{
Esta funci?n calcula la varianza de un vector num?rico. La varianza es una medida de la dispersi?n o variabilidad de los datos.
}
\examples{
Ejemplo de uso
datos <- c(1, 2, 3, 4, 5)
var <- calcular_varianza(datos) # Calcula la varianza del vector
var

#Ejemplo con un solo valor
valor <- 7
calcular_varianza(valor)  # Devuelve 0, ya que la varianza de un solo valor es 0.

Ejemplo con un vector de longitud 1
vector_pequeno <- c(5)
calcular_varianza(vector_pequeno)  # Devuelve 0, ya que la varianza de un vector de #'un solo valor es 0.

}
\seealso{
Puede usar esta funci?n para calcular la varianza de atributos num?ricos en un conjunto de datos.  Se la llama desde la funci?n 'calcular_metricas'.
}
