% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Exportacion_Carga.R
\name{cargar_datos_desde_csv}
\alias{cargar_datos_desde_csv}
\title{Cargar datos desde un archivo CSV utilizando la biblioteca readr}
\usage{
cargar_datos_desde_csv(nombre_archivo_csv)
}
\arguments{
\item{nombre_archivo_csv}{El nombre del archivo CSV desde el cual se cargar?n los datos.}
}
\value{
Un DataFrame que contiene los datos cargados desde el archivo CSV, o NULL si no se pudieron cargar los datos.
}
\description{
Esta funci?n carga datos desde un archivo CSV utilizando la biblioteca readr. Se espera que el archivo CSV tenga un formato v?lido y contenga datos tabulares. Los datos cargados se almacenan en un DataFrame y se muestran en la consola.
}
\examples{
Ejemplo de uso
Llamar a la funci?n para cargar datos desde un archivo CSV
datos_cargados <- cargar_datos_desde_csv(nombre_archivo_csv = "resultados.csv")

}
