% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Discretizacion.R
\name{Discretizacion_Basada_En_Clusteres_Simple}
\alias{Discretizacion_Basada_En_Clusteres_Simple}
\title{Discretizaci?n basada en Cl?steres, m?todo K-Means}
\usage{
Discretizacion_Basada_En_Clusteres_Simple(atributo, n_clusters, plot = TRUE)
}
\arguments{
\item{atributo}{Un vector num?rico que se va a discretizar.}

\item{n_clusters}{El n?mero de cl?steres en los que se va a discretizar el atributo.}

\item{plot}{Un valor l?gico que indica si se deben mostrar gr?ficos de los resultados (TRUE o FALSE).}
}
\value{
Un vector de etiquetas que indica a qu? cl?ster pertenece cada valor del atributo.Y el plot si asi se ha indicado.
}
\description{
Esta funci?n discretiza un atributo num?rico utilizando el algoritmo K-Means con una implementaci?n simple y opcionalmente muestra gr?ficos de los resultados.
}
\examples{

Ejemplo de uso
datos <- data.frame(
 Valores = c(10, 15, 600, 20, 35, 80, 85, 55, 60, 300, 400, 500, 80, 90, 90)
)
n_clusters <- 5

Discretizado_Clusteres_Simple <- Discretizacion_Basada_En_Clusteres_Simple(datos$Valores, n_clusters, plot = TRUE)
print(Discretizado_Clusteres_Simple)

}
