% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Discretizacion.R
\name{Discretizacion_dataset}
\alias{Discretizacion_dataset}
\title{Discretizaci?n de un dataset completo}
\usage{
Discretizacion_dataset(dataset, n_intervalos, metodo, plot)
}
\arguments{
\item{dataset}{El dataset a discretizar.}

\item{n_intervalos}{El n?mero de intervalos en los que se va a discretizar cada variable.}

\item{metodo}{El m?todo de discretizaci?n a utilizar. Puede ser "anchura" o "frecuencia".}

\item{plot}{Un valor l?gico que indica si se deben mostrar gr?ficos de los resultados (TRUE o FALSE).}
}
\value{
Una lista que contiene las variables del dataset discretizadas de acuerdo al m?todo especificado.
}
\description{
Esta funci?n discretiza todas las variables num?ricas en un dataset en intervalos de igual anchura o igual frecuencia, seg?n el m?todo especificado, y opcionalmente muestra gr?ficos de los resultados.
}
\examples{

Ejemplo de uso
datos <- data.frame(
 'Edad' = c(30, 40, 35, 28, 45, 37, 32, 41, 29, 33),
 'Ingresos' = c(55000, 48000, 60000, 52000, 58000, 49000, 57000, 51000, 54000, 63000),
 'Puntuacion' = c(75, 68, 72, 80, 67, 73, 78, 69, 74, 76)
)
n_intervalos <- 5

dataset_discretizado_anchura <- Discretizacion_dataset(datos, n_intervalos, metodo = "anchura", plot= TRUE)
dataset_discretizado_anchura
dataset_discretizado_frecuencia <- Discretizacion_dataset(datos, n_intervalos, metodo = "frecuencia", plot= TRUE)
dataset_discretizado_frecuencia

Ejemplo de uso con valores NA
datos <- data.frame(
 'Edad' = c(30, 40, NA, 28, 45, 37, 32, 41, 29, 33),
 'Ingresos' = c(55000, NA, 60000, NA, 58000, 49000, 57000, 51000, 54000, 63000),
 'Puntuacion' = c(75, 68, 72, 80, 67, 73, 78, 69, 74, 76)
)
n_intervalos <- 5
dataset_discretizado_anchura <- Discretizacion_dataset(datos, n_intervalos, metodo = "anchura", plot= TRUE)
dataset_discretizado_anchura
dataset_discretizado_frecuencia <- Discretizacion_dataset(datos, n_intervalos, metodo = "frecuencia", plot= TRUE)
dataset_discretizado_frecuencia
}
