% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Discretizacion.R
\name{Discretizacion_igual_anchura}
\alias{Discretizacion_igual_anchura}
\title{Discretizaci?n por igual anchura}
\usage{
Discretizacion_igual_anchura(datos, atributo_a_discretizar, n_intervalos)
}
\arguments{
\item{datos}{Un dataframe que contiene los datos a procesar.}

\item{atributo_a_discretizar}{El nombre del atributo que se va a discretizar.}

\item{n_intervalos}{El n?mero de intervalos en los que se va a discretizar el atributo.}
}
\value{
Un vector de enteros que representa la discretizaci?n del atributo.
}
\description{
Esta funci?n discretiza un atributo en un conjunto de datos en intervalos de igual anchura.
}
\seealso{
Puede usar esta funci?n para calcular la discretizaci?n por igual anchura en un conjunto de datos.  Se la llama desde la funci?n 'Discretizacion_dataset'.
}
