% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Discretizacion.R
\name{Discretizacion_igual_frecuencia}
\alias{Discretizacion_igual_frecuencia}
\title{Discretizaci?n por igual frecuencia}
\usage{
Discretizacion_igual_frecuencia(datos, atributo_a_discretizar, n_intervalos)
}
\arguments{
\item{datos}{Un dataframe que contiene los datos a procesar.}

\item{atributo_a_discretizar}{El nombre del atributo que se va a discretizar.}

\item{n_intervalos}{El n?mero de intervalos en los que se va a discretizar el atributo.}
}
\value{
Un vector de enteros que representa la discretizaci?n del atributo.
}
\description{
Esta funci?n discretiza un atributo en un conjunto de datos en intervalos de igual frecuencia.
}
\examples{
Ejemplo de uso
datos <- data.frame(
'Edad' = c(30, 40, 35, 28, 45, 37, 32, 41, 29, 33),
'Ingresos' = c(55000, 48000, 60000, 52000, 58000, 49000, 57000, 51000, 54000, 63000),
'Puntuacion' = c(75, 68, 72, 80, 67, 73, 78, 69, 74, 76)
)
n_intervalos <- 5
discretizacion <- Discretizacion_igual_frecuencia(datos, "Edad", n_intervalos)

}
