% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Normalizar_estandarizar.R
\name{estandarizar_vector}
\alias{estandarizar_vector}
\title{Funci?n para estandarizar un vector num?rico (media 0, desviaci?n est?ndar 1)}
\usage{
estandarizar_vector(vector_numerico)
}
\arguments{
\item{vector_numerico}{Un vector num?rico que se desea estandarizar.}
}
\value{
Un vector num?rico estandarizado con media 0 y desviaci?n est?ndar 1. Si no es posible estandarizar el vector (por falta de variaci?n), se muestra un mensaje
de advertencia y se devuelve el vector original sin cambios.
}
\description{
Esta funci?n estandariza un vector num?rico para que tenga una media de 0 y una desviaci?n est?ndar de 1. La estandarizaci?n se realiza restando la media y dividiendo por la desviaci?n est?ndar del vector.
}
\examples{
Ejemplo de uso
datos <- c(10, 20, 30, 40, 50)
datos_estandarizados <- estandarizar_vector(datos) # Estandariza el vector con media 0 y desviaci?n est?ndar 1.
datos_estandarizados

Ejemplo con falta de variaci?n
datos_faltavariacion <- c(10, 10, 10, 10, 10)
estandarizar_vector(datos_faltavariacion)  # Muestra un mensaje de advertencia y devuelve el vector original.

}
