% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Exportacion_Carga.R
\name{exportar_a_CSV}
\alias{exportar_a_CSV}
\title{Exportar un DataFrame a un archivo CSV y mostrar la ruta del archivo creado}
\usage{
exportar_a_CSV(df, nombre_archivo_csv)
}
\arguments{
\item{df}{El DataFrame que se exportar? a un archivo CSV.}

\item{nombre_archivo_csv}{El nombre del archivo CSV en el que se guardar?n los datos.}
}
\value{
El contenido del archivo CSV creado.
}
\description{
Esta funci?n toma un DataFrame, lo convierte en un archivo CSV y muestra la ruta absoluta del archivo creado. Es ?til para guardar y compartir conjuntos de datos en formato CSV.
}
\examples{
Ejemplo de uso
datos<- data.frame(
 Edad = c(25, 30, 35, 40, 45),
 Altura = c(160, 175, 180, 170, 185),
 Peso = c(60, 70, 75, 80, 90)
)
Nombra el archivo CSV en el que deseas guardar los datos y llama a la funci?n para exportar
exportar_a_CSV(datos, nombre_archivo_csv="resultados.csv")
}
