% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Filtrado.R
\name{filtrar_variables_por_entropia}
\alias{filtrar_variables_por_entropia}
\title{Filtrar variables de un dataset basado en la entrop?a}
\usage{
filtrar_variables_por_entropia(dataset, umbral_entropia)
}
\arguments{
\item{dataset}{El dataset del cual se van a filtrar las variables.}

\item{umbral_entropia}{El umbral de entrop?a. Las variables con una entrop?a mayor que este umbral se mantienen en el dataset resultante.}
}
\value{
Un nuevo dataset que contiene las variables que cumplen con el criterio de entrop?a especificado.
}
\description{
Esta funci?n filtra las variables de un dataset basado en un umbral de entrop?a. Las variables cuyas entrop?as superen el umbral se mantienen en el dataset resultante.
}
\examples{
Ejemplo de uso
dataset <- list(
 Variable1 = c("A", "A", "A", "C", "B"),
 Variable2 = c("X", "Y", "X", "Z", "Z"),
 Variable3 = c("A", "B", "A", "C", "B")
)
Definir un umbral de entrop?a
umbral_entropia <- 1.4
Filtrar las variables basadas en la entrop?a
variables_filtradas <- filtrar_variables_por_entropia(dataset, umbral_entropia)
print(variables_filtradas)

}
