% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Normalizar_estandarizar.R
\name{normalizar_estandarizar_dataset}
\alias{normalizar_estandarizar_dataset}
\title{Funci?n para normalizar y estandarizar un dataset num?rico}
\usage{
normalizar_estandarizar_dataset(dataset)
}
\arguments{
\item{dataset}{El dataset que se desea normalizar y estandarizar}
}
\value{
Una lista que contiene dos versiones del dataset: uno normalizado y otro estandarizado. Cada versi?n tiene las mismas columnas que el dataset original, pero con los valores normalizados  o estandarizados, respectivamente.
}
\description{
Esta funci?n toma un conjunto de datos y aplica normalizaci?n y estandarizaci?n a todas las columnas num?ricas del conjunto de datos. La normalizaci?n escala los valores de cada columna a un rango de 0 a 1, mientras que la estandarizaci?n ajusta la media a 0 y la desviaci?n est?ndar a 1 en cada columna num?rica.
}
\examples{
# Crear un conjunto de datos de ejemplo
datos <- data.frame(
  Atributo1 = c(10, 20, 30, 40, 50),
  Atributo2 = c(1, 2, 3, 4, 5)
)

# Aplicar normalizaci?n y estandarizaci?n al conjunto de datos de ejemplo
resultados <- normalizar_estandarizar_dataset(datos)
normalizado <- resultados$normalizado # Dataset normalizado
estandarizado <- resultados$estandarizado # Dataset estandarizado
normalizado
estandarizado


}
