% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Normalizar_estandarizar.R
\name{normalizar_vector}
\alias{normalizar_vector}
\title{Funci?n para normalizar un vector num?rico en una escala de 0 a 1}
\usage{
normalizar_vector(vector_numerico)
}
\arguments{
\item{vector_numerico}{Un vector num?rico que se desea normalizar.}
}
\value{
Un vector num?rico normalizado en una escala de 0 a 1.
}
\description{
Esta funci?n normaliza un vector num?rico en una escala de 0 a 1. La normalizaci?n se realiza dividiendo cada valor por la diferencia entre el valor m?ximo y m?nimo del vector.
}
\examples{
Ejemplo de uso
datos <- c(10, 20, 30, 40, 50)
datos_normalizados <- normalizar_vector(datos)  # Normaliza el vector en una escala de 0 a 1.
datos_normalizados

Ejemplo con falta de variaci?n
datos_faltavariacion <- c(10, 10, 10, 10, 10)
normalizar_vector(datos_faltavariacion)  # Muestra un mensaje de advertencia y devuelve el vector original.


}
