% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_PLOT.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Funci?n para calcular la matriz de correlacion y trazar la matriz utilizando el metodo especificado (correlacion o informacion mutua)}
\usage{
plot_correlation_matrix(dataset, method)
}
\arguments{
\item{dataset}{El conjunto de datos utilizado para calcular la matriz de correlaci?n.}

\item{method}{En m?todo utilizado para calcular la matriz de correlaci?n. Puede ser 'correlacion' o 'mutua}
}
\value{
Un gr?fico de la Curva ROC y el valor del ?rea Bajo la Curva ROC (AUC).
}
\description{
Esta funci?n calcula la matriz de correlacion y la representa segun el metodo seleccionado.
}
\examples{
Ejemplo de uso
dataset <- data.frame(
 Variable1 = c(1.8, 2.6, 4.5, 4.0, 1.2),
 Variable2 = c(0.7, 2.0, 3.5, 6.0, 0.5),
 Variable3 = factor(c("B", "B", "A", "C", "C")),
 Variable4 = factor(c("B", "Y", "A", "Z", "Z"))
)
plot_correlation_matrix (dataset, 'correlacion')
plot_correlation_matrix (dataset, 'mutua')

}
