% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_PLOT.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Funci?n para calcular la matriz de correlacion y trazar la matriz utilizando el metodo especificado (correlacion o informacion mutua)}
\usage{
plot_matrix(correlation_matrix, title)
}
\arguments{
\item{correlation_matrix}{El conjunto de datos}

\item{title}{En m?todo utilizado para calcular la matriz de correlaci?n. Puede ser 'correlacion' o 'mutua}
}
\value{
Un gr?fico de la Curva ROC y el valor del ?rea Bajo la Curva ROC (AUC).
}
\description{
Representacion de la matriz
}
